requirements.txt
python-telegram-bot
gmusicapi
youtube-dl
pafy
pydub
simpleaudio
pylru
soundcloud
colorama
hug
passlib
pyjwt
bcrypt
aiohttp_wsgi
